Move sqlitedb tests to own file
This change moves the sqlitedb integration tests to their own file.
This should improve readability of the submodules.
